Move Polygon object from DTO to VO.
