+++ src/Service/CoordinatesService.php
@@ -4,7 +4,7 @@
 namespace App\Service;
 
 
-use App\Dto\PolygonDTO;
+use App\Vo\PolygonVO;
 use App\Entity\Place;
 use Doctrine\Common\Collections\Collection;
 use Exception;
@@ -14,11 +14,11 @@ class CoordinatesService
     /**
      * @param Collection<int, Place> $places
      *
-     * @return PolygonDTO
+     * @return PolygonVO
      *
      * @throws Exception
      */
-    public function getPolygon(Collection $places): PolygonDTO
+    public function getPolygon(Collection $places): PolygonVO
     {
         $places = $places->toArray();
 
@@ -52,7 +52,7 @@ class CoordinatesService
         $highestLon = $places[0]->getLon();
         $lowerLon = $places[array_key_last($places)]->getLon();
 
-        $polygon = new PolygonDTO($highestLat, $highestLon, $lowerLat, $lowerLon);
+        $polygon = new PolygonVO($highestLat, $highestLon, $lowerLat, $lowerLon);
 
         return $polygon;
     }
